% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{add_all_possible_shims}
\alias{add_all_possible_shims}
\title{Write shims for all possible shimmable functions}
\usage{
add_all_possible_shims()
}
\description{
Write shims for all possible shimmable functions
}
