% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{add_shim}
\alias{add_shim}
\title{Add a function to the list of functions that get checked for file path issues.}
\usage{
add_shim(func, package = "", path_arg = "")
}
\arguments{
\item{func}{name of function you want to create a shim for (e.g. "\code{read_excel}")}

\item{package}{name of package that provided function is from (e.g. "\code{readxl}")}

\item{path_arg}{name of path-related argument in that function (if not
specified, fertile will make an educated guess).}
}
\description{
Add a function to the list of functions that get checked for file path issues.
}
