% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{check}
\alias{check}
\alias{print.fertile}
\title{Run reproducibility checks}
\usage{
check(path = ".")

\method{print}{fertile}(x, ...)
}
\arguments{
\item{path}{Path to package root}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{\link{fertile}} object
}
\description{
Run reproducibility checks
}
\examples{
check()
}
