% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_from_zip}
\alias{check_from_zip}
\title{Test projects in ZIP files}
\usage{
check_from_zip(url, ...)
}
\arguments{
\item{url}{The URL to download.}

\item{...}{Other arguments that are passed to \code{\link{download.file}}.}
}
\description{
Test projects in ZIP files
}
