% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_is_dir}
\alias{check_is_dir}
\title{Utility function to check whether a provided path is a directory}
\usage{
check_is_dir(path)
}
\description{
Utility function to check whether a provided path is a directory
}
