% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_is_file}
\alias{check_is_file}
\title{Check whether a provided path is a file}
\usage{
check_is_file(path)
}
\arguments{
\item{path}{Path you are wanting to check}
}
\description{
Check whether a provided path is a file
}
\seealso{
Other path type checks: 
\code{\link{check_is_dir}()}
}
\concept{path type checks}
