% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{check_path}
\alias{check_path}
\title{Check Paths for portability}
\usage{
check_path(path, strict = TRUE)
}
\arguments{
\item{path}{a vector of paths}

\item{strict}{logical indicating whether you want to stop on errors}
}
\description{
Check Paths for portability
}
\examples{
check_path(tempfile())
check_path(tempfile(), strict = FALSE)
check_path(c("data.csv", "~/.Rprofile"), strict = FALSE)
}
