% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{check_path_safe}
\alias{check_path_safe}
\title{Check paths ONLY when running in interactive mode}
\usage{
check_path_safe(path)
}
\arguments{
\item{path}{path you want to check}
}
\description{
Check paths ONLY when running in interactive mode
}
