% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{check_path_shim}
\alias{check_path_shim}
\title{Check paths for portability}
\usage{
check_path_shim(path, parent = ".", strict = TRUE, ...)
}
\arguments{
\item{path}{a vector of paths}

\item{strict}{logical indicating whether you want to stop on errors}

\item{...}{For behind the scenes behavior by fertile. Please ignore.}
}
\description{
Check paths for a variety of maladies
}
\keyword{internal}
