% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/fertile.R
\name{is_readme_exists}
\alias{is_readme_exists}
\alias{has_proj_root}
\alias{has_no_absolute_paths}
\alias{has_only_portable_paths}
\alias{has_no_randomness}
\alias{checks}
\alias{check}
\title{Reproducbility checks}
\usage{
is_readme_exists(path = ".", ...)

has_proj_root(path = ".")

has_no_absolute_paths(path = ".")

has_only_portable_paths(path = ".")

has_no_randomness(path = ".")

check(path = ".", ...)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{...}{currently ignore}
}
\description{
Reproducbility checks
}
