% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{danger}
\alias{danger}
\title{Danger block}
\usage{
danger(expr)
}
\arguments{
\item{expr}{Code to run as if \code{fertile} was not loaded}
}
\description{
Danger block
}
\examples{
danger(setwd(tempdir()))
}
