% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{disable_added_shims}
\alias{disable_added_shims}
\title{Remove all user-added shims from the global environment}
\usage{
disable_added_shims()
}
\description{
Remove all user-added shims from the global environment
}
