% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onAttach}
\alias{.onAttach}
\title{Load shims into environment when fertile is attached}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where the package defining the namespace was found}

\item{pkgname}{a character string giving the name of the package}
}
\description{
Load shims into environment when fertile is attached
}
