% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onDetach}
\alias{.onDetach}
\title{Remove shims from environment when fertile is detached}
\usage{
.onDetach(libpath)
}
\arguments{
\item{libpath}{a character string giving the complete path to the package}
}
\description{
Remove shims from environment when fertile is detached
}
