% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{enable_added_shims}
\alias{enable_added_shims}
\title{Remove all user-added shims from the global environment}
\usage{
enable_added_shims()
}
\description{
Remove all user-added shims from the global environment
}
