% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_exists_within}
\alias{file_exists_within}
\title{Does file exist in project tree?}
\usage{
file_exists_within(path)
}
\arguments{
\item{path}{A vector of paths}
}
\value{
A logical vector
}
\description{
Does file exist in project tree?
}
\examples{
file_exists_within(test_paths$path)
}
