% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_all_shimmable_functions}
\alias{find_all_shimmable_functions}
\title{Find the names of all shimmable functions within the list of loaded packages}
\usage{
find_all_shimmable_functions()
}
\value{
list containing the names of all the shimmable functions and their associated packages
}
\description{
Find the names of all shimmable functions within the list of loaded packages
}
\keyword{internal}
