% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_pkg_shimmable_functions}
\alias{find_pkg_shimmable_functions}
\title{Find the names of all functions that are potentially shimmable for a given package}
\usage{
find_pkg_shimmable_functions(package)
}
\arguments{
\item{package}{name of package to search through}
}
\value{
vector containing the names of all the shimmable functions for the provided package
}
\description{
Find the names of all functions that are potentially shimmable for a given package
}
\keyword{internal}
