% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_name}
\alias{function_name}
\title{Sub-function to help disable_added_shims operate: #2}
\usage{
function_name(expr)
}
\description{
Sub-function to help disable_added_shims operate: #2
}
\keyword{internal}
