% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{generate_script}
\alias{generate_script}
\title{Utility function for proj_pkg_script}
\usage{
generate_script(pkg_name, vector = c())
}
\arguments{
\item{pkg_name}{Name of package to generate install script for}
}
\description{
Utility function for proj_pkg_script
}
\keyword{internal}
