% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_shim_code}
\alias{get_shim_code}
\title{Generate the code associated with writing a shim}
\usage{
get_shim_code(func, package = "", path_arg = "")
}
\arguments{
\item{func}{name of function you want to create a shim for (e.g. "read_excel")}

\item{package}{name of package that provided function is from (e.g. "readxl")}

\item{path_arg}{name of path-related argument in that function (if not specified, fertile will make an educated guess).}
}
\value{
vector of lines making up code for shim
}
\description{
Generate the code associated with writing a shim
}
\keyword{internal}
