% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_rendered}
\alias{has_rendered}
\title{Utility function to check whether a project has been updated since last rendered}
\usage{
has_rendered(path = ".")
}
\arguments{
\item{path}{Path to the project}
}
\description{
Utility function to check whether a project has been updated since last rendered
}
