% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{interactive_log_on}
\alias{interactive_log_on}
\title{Utility function to help w/ controlling interactive logging functionality}
\usage{
interactive_log_on()
}
\description{
Utility function to help w/ controlling interactive logging functionality
}
