% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_data_file}
\alias{is_data_file}
\title{Test whether a given path is to a data file}
\usage{
is_data_file(path)
}
\arguments{
\item{path}{Path to file you want to test}
}
\description{
Test whether a given path is to a data file
}
