% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_function}
\alias{is_function}
\title{Sub-function to help disable_added_shims operate: #1}
\usage{
is_function(expr)
}
\description{
Sub-function to help disable_added_shims operate: #1
}
\keyword{internal}
