% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_text_file}
\alias{is_text_file}
\title{Test whether a given path is to a text file}
\usage{
is_text_file(path)
}
\arguments{
\item{path}{Path to file you want to test}
}
\description{
Test whether a given path is to a text file
}
\seealso{
Other file type checks: \code{\link{is_data_file}},
  \code{\link{is_image_file}}, \code{\link{is_r_file}}
}
\concept{file type checks}
