% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_report}
\alias{log_report}
\alias{render_log_report}
\alias{log_clear}
\alias{path_log}
\title{Logging and reporting of file paths}
\usage{
log_report(path = proj_root())

render_log_report(path = proj_root())

log_clear(path = proj_root())

path_log(path = proj_root())
}
\arguments{
\item{path}{Path to the directory you want to access the log for}
}
\description{
These functions provide access to a log file, located in the
directory given by path `x`, that records the executed paths/arguments passed to
commonly-used input and output functions in R code located in that directory.
}
\section{log_report}{

Return the log of functions/paths you have used interactively while
within a project directory.

\code{log_report("directory")}
OR, to look at your project root:
\code{log_report()}
}

\section{render_log_report}{

View the render log for a directory, which shows information about
paths, packages, and randomness seeds involved in the rendering
of a project's files.
\code{render_log_report("directory")}
}

\section{log_clear}{

Delete/clear the log file for a directory.

\code{log_clear("directory")}
OR, to clear your project root:
\code{log_clear()}
}

\section{path_log}{

Return the path to the log file in a given directory.

\code{path_log("directory")}
OR, to get the log file for your project root:
\code{path_log()}
}

