% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_touch}
\alias{log_touch}
\title{Create a new log file and return the path to the log}
\usage{
log_touch(path = proj_root())
}
\description{
Create a new log file and return the path to the log
}
\keyword{internal}
