% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{make_check}
\alias{make_check}
\title{Create a fertile check object}
\usage{
make_check(fun, name, req_compilation, state, problem, solution, help,
  errors, ...)
}
\arguments{
\item{fun}{Function to run that returns a logical}

\item{name}{What is the check checking for?}

\item{req_compilation}{Does the code have to be compiled for the check to
work?}

\item{state}{Logical indicating whether the check passed/failed}

\item{problem}{Description of the problem}

\item{solution}{Description of a potential solution}

\item{help}{References to help files}

\item{errors}{A \code{\link[tibble]{tibble}} of error messages}
}
\description{
Create a fertile check object
}
\seealso{
\code{\link[usethis]{use_tidy_style}},
         \code{\link[fs]{is_absolute_path}}
}
\keyword{internal}
