% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_check}
\alias{path_check}
\title{Check Paths for portability}
\usage{
path_check(path)
}
\arguments{
\item{path}{a vector of paths}
}
\description{
Check Paths for portability
}
\examples{
path_check(tempfile())
}
