% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_within}
\alias{path_within}
\title{Check if paths ore relative to project root}
\usage{
path_within(path, start = ".")
}
\arguments{
\item{path}{a vector of paths}
}
\value{
a logical vector
}
\description{
Check if paths ore relative to project root
}
\examples{
path_within(test_paths$path)
}
