% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{print.fertile_check}
\alias{print.fertile_check}
\title{Print method for fertile checks}
\usage{
\method{print}{fertile_check}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for fertile checks
}
