% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{proj_badges}
\alias{proj_badges}
\title{Get reproducibility badges for a project}
\usage{
proj_badges(path = ".")
}
\arguments{
\item{path}{Path to project root}
}
\value{
Path to an html output file summarizing the badges received/failed
}
\description{
Get reproducibility badges for a project
}
