% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{proj_dependency_report}
\alias{proj_dependency_report}
\title{Return a sessionInfo() style report of the packages/software versions
that your R project was last run successfully on.}
\usage{
proj_dependency_report(path = proj_root())
}
\arguments{
\item{path}{path to the project directory}
}
\description{
Return a sessionInfo() style report of the packages/software versions
that your R project was last run successfully on.
}
