% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{proj_pkg_script}
\alias{proj_pkg_script}
\title{Generate an R script to install all of the packages
required to run the R/Rmd files in an R project.
Once generated, the script can be found in the root
directory of the project.}
\usage{
proj_pkg_script(path = ".")
}
\arguments{
\item{path}{Path to project root}
}
\value{
An R script file ("install_proj_packages.r")
}
\description{
Generate an R script to install all of the packages
required to run the R/Rmd files in an R project.
Once generated, the script can be found in the root
directory of the project.
}
