% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{proj_render}
\alias{proj_render}
\title{Render files in a project directory to update the render log file}
\usage{
proj_render(path = ".", ...)
}
\arguments{
\item{path}{Path to project root}
}
\description{
Render files in a project directory to update the render log file
}
\keyword{internal}
