% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_root}
\alias{proj_root}
\title{Find the project root, but always return something}
\usage{
proj_root(path = ".")
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Find the project root, but always return something
}
