% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{read_csv}
\alias{read_csv}
\title{Shims for common input/output functions}
\usage{
read_csv(file, ...)
}
\arguments{
\item{file}{The path}

\item{...}{arguments passed to functions}
}
\description{
Shims for common input/output functions
}
\examples{
if (require(readr)) {
  read_csv(test_paths$path)
}
}
\seealso{
\code{\link[readr]{read_csv}}
}
