% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements.R
\name{req_code}
\alias{req_code}
\title{Extract requirements from code}
\usage{
req_code(x)
}
\arguments{
\item{x}{Code to examine. Supports unquoting.}
}
\description{
Looks for `::`, `:::`, `library()`, `require()`, `requireNamespace()`,
and `loadNamespace()`.
}
\keyword{internal}
