% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{takes_path_arg}
\alias{takes_path_arg}
\title{Return name of a given function's file path-related argument}
\usage{
takes_path_arg(func, package = "")
}
\arguments{
\item{func}{name of function to check for path argument (e.g. "read_csv")}

\item{package}{name of package that provided function is from (e.g. "readr")}
}
\description{
Return name of a given function's file path-related argument
}
\keyword{internal}
