% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_paths}
\alias{test_paths}
\title{Sample paths to test}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 5 columns.}
\usage{
test_paths
}
\description{
Sample paths to test
}
\keyword{datasets}
