% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_execute}
\alias{to_execute}
\title{load a list of packages and export the session info of them}
\usage{
to_execute(packages_to_load, path)
}
\arguments{
\item{packages_to_load}{vector of package names}

\item{path}{path of directory we are working in}
}
\description{
load a list of packages and export the session info of them
}
\keyword{internal}
