% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R, R/utils.R
\name{add_shim}
\alias{add_shim}
\alias{edit_shims}
\alias{unload_shims}
\alias{load_shims}
\alias{add_all_possible_shims}
\alias{path_shims}
\alias{read_shims}
\alias{active_shims}
\title{Add a function to the list of functions that get checked for file path issues.}
\usage{
add_shim(func, package = "", path_arg = "")

edit_shims()

unload_shims()

load_shims()

add_all_possible_shims()

path_shims()

read_shims()

active_shims()
}
\arguments{
\item{func}{name of function you want to create a shim for (e.g. "\code{read_excel}")}

\item{package}{name of package that provided function is from (e.g. "\code{readxl}")}

\item{path_arg}{name of path-related argument in that function (if not
specified, fertile will make an educated guess).}
}
\description{
Add a function to the list of functions that get checked for file path issues.

View/edit list of user created shims.

Remove all user-added shims from the global environment

Write shims for all possible shimmable functions

Check that shims file exists and return path

Read shim file

Get list of active shims in global environment
}
