% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/fertile.R
\name{has_tidy_media}
\alias{has_tidy_media}
\alias{has_tidy_images}
\alias{has_tidy_code}
\alias{has_tidy_raw_data}
\alias{has_tidy_data}
\alias{has_tidy_scripts}
\alias{has_readme}
\alias{has_proj_root}
\alias{has_no_absolute_paths}
\alias{has_only_portable_paths}
\alias{has_no_randomness}
\alias{check}
\alias{print.fertile_check}
\title{Reproducbility checks}
\usage{
has_tidy_media(path = ".", ...)

has_tidy_images(path = ".", ...)

has_tidy_code(path = ".", ...)

has_tidy_raw_data(path = ".", ...)

has_tidy_data(path = ".", ...)

has_tidy_scripts(path = ".", ...)

has_readme(path = ".", ...)

has_proj_root(path = ".", ...)

has_no_absolute_paths(path = ".", ...)

has_only_portable_paths(path = ".", ...)

has_no_randomness(path = ".", seed_old, ...)

check(path = ".", ...)

\method{print}{fertile_check}(x, ...)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{...}{currently ignored}

\item{seed_old}{The old seed before the code is rendered}

\item{x}{an object used to select a method.}
}
\value{
a \code{\link[tibble]{tibble}} of checks and their results
}
\description{
A laundry list of small checks that help make your project
more likely to be reproducible
}
