% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{check}
\alias{check}
\alias{print.fertile}
\title{Run reproducibility checks}
\usage{
check(path = ".")

\method{print}{fertile}(x, ...)
}
\arguments{
\item{path}{Path to package root}
}
\value{
A \code{\link{fertile}} object
}
\description{
Run reproducibility checks
}
\examples{
check()
}
