% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/fertile.R
\name{make_check}
\alias{make_check}
\alias{print.fertile_check}
\alias{print_one_check}
\alias{has_tidy_media}
\alias{has_tidy_images}
\alias{has_tidy_code}
\alias{has_tidy_raw_data}
\alias{has_tidy_data}
\alias{has_tidy_scripts}
\alias{has_readme}
\alias{has_proj_root}
\alias{has_no_nested_proj_root}
\alias{has_no_absolute_paths}
\alias{has_only_portable_paths}
\alias{has_no_randomness}
\alias{has_no_lint}
\alias{has_clear_build_chain}
\alias{check}
\title{Reproducbility checks}
\usage{
make_check(fun, name, req_compilation, state, problem, solution, help,
  errors, ...)

\method{print}{fertile_check}(x, ...)

print_one_check(x, ...)

has_tidy_media(path = ".", ...)

has_tidy_images(path = ".", ...)

has_tidy_code(path = ".", ...)

has_tidy_raw_data(path = ".", ...)

has_tidy_data(path = ".", ...)

has_tidy_scripts(path = ".", ...)

has_readme(path = ".", ...)

has_proj_root(path = ".", ...)

has_no_nested_proj_root(path = ".", ...)

has_no_absolute_paths(path = ".", ...)

has_only_portable_paths(path = ".", ...)

has_no_randomness(path = ".", seed_old, ...)

has_no_lint(path = ".", ...)

has_clear_build_chain(path = ".", ...)

check(path = ".", ...)
}
\arguments{
\item{fun}{Function to run that returns a logical}

\item{name}{What is the check checking for?}

\item{req_compilation}{Does the code have to be compiled for the check to
work?}

\item{state}{Logical indicating whether the check passed/failed}

\item{problem}{Description of the problem}

\item{solution}{Description of a potential solution}

\item{help}{References to help files}

\item{errors}{A \code{\link[tibble]{tibble}} of error messages}

\item{...}{currently ignored}

\item{x}{an object used to select a method.}

\item{path}{A character vector of one or more paths.}

\item{seed_old}{The old seed before the code is rendered}
}
\value{
a \code{\link[tibble]{tibble}} of checks and their results
}
\description{
A laundry list of small checks that help make your project
more likely to be reproducible
}
\seealso{
\code{\link[usethis]{use_tidy_style}},
         \code{\link[fs]{is_absolute_path}}
}
