% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_is_dir}
\alias{check_is_dir}
\title{Check whether a provided path is a directory}
\usage{
check_is_dir(path)
}
\arguments{
\item{path}{Path you are wanting to check}
}
\description{
Check whether a provided path is a directory
}
\seealso{
Other path type checks: \code{\link{check_is_file}}
}
\concept{path type checks}
