% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_is_file}
\alias{check_is_file}
\title{Utility function to check whether a provided path is a file}
\usage{
check_is_file(path)
}
\arguments{
\item{path}{Path you are wanting to check}
}
\description{
Utility function to check whether a provided path is a file
}
