% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{is_path_portable}
\alias{is_path_portable}
\alias{check_path_is_portable}
\alias{check_path_absolute}
\alias{check_path}
\title{Check paths for portability}
\usage{
is_path_portable(path, parent = ".")

check_path_is_portable(path, parent = ".", strict = TRUE)

check_path_absolute(path, strict = TRUE)

check_path(path, parent = ".", strict = TRUE)
}
\arguments{
\item{path}{a vector of paths}

\item{parent}{The parent path.}

\item{strict}{logical indicating whether you want to stop on errors}
}
\value{
A logical vector
}
\description{
Check paths for a variety of maladies
}
\examples{
is_path_portable(c(tempfile(), "~/.Rprofile", "../data.csv"))
\dontrun{
check_path(tempfile())
}
check_path(tempfile(), strict = FALSE)
check_path(c("data.csv", "~/.Rprofile"), strict = FALSE)
}
