% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R, R/paths.R
\name{check_file_exists}
\alias{check_file_exists}
\alias{is_path_here}
\alias{check_path_here}
\alias{path_rel_here}
\alias{check_path}
\title{Check paths for portability}
\usage{
check_file_exists(path, strict = TRUE)

is_path_here(path)

check_path_here(path, strict = TRUE)

path_rel_here(path)

check_path(path, strict = TRUE)
}
\arguments{
\item{path}{a vector of paths}

\item{strict}{logical indicating whether you want to stop on errors}
}
\value{
A logical vector
}
\description{
Check paths for a variety of maladies
}
\examples{
\dontrun{
check_file_exists(tempfile())
}
is_path_here(tempfile())
\dontrun{
check_path(tempfile())
}
check_path(tempfile(), strict = FALSE)
check_path(c("data.csv", "~/.Rprofile"), strict = FALSE)
}
