% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{check_path_safe}
\alias{check_path_safe}
\title{Runs check_path, but will only work interactively and cannot be run by fertile in the background.}
\usage{
check_path_safe(path, ...)
}
\arguments{
\item{path}{Path you want to check}
}
\description{
Runs check_path, but will only work interactively and cannot be run by fertile in the background.
}
\keyword{internal}
