% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R, R/files.R, R/paths.R
\name{checks}
\alias{checks}
\alias{file_exists_here}
\alias{check_file_here}
\alias{is_path_here}
\alias{check_path_here}
\title{File checks}
\usage{
checks(file)

file_exists_here(path)

check_file_here(path)

is_path_here(path)

check_path_here(path)
}
\arguments{
\item{file}{a path to a file}

\item{path}{A vector of paths}
}
\value{
A logical vector

A logical vector
}
\description{
File checks

Does file exist in project tree?

Check if paths are relative to project root
}
\examples{
checks(test_paths$path[1])
checks(test_paths$path[3])
file_exists_here(test_paths$path)
check_file_here("data.csv")
is_path_here(test_paths$path)
}
