% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/fertile.R
\name{has_readme}
\alias{has_readme}
\alias{has_proj_root}
\alias{has_no_absolute_paths}
\alias{has_only_portable_paths}
\alias{has_no_randomness}
\alias{checks}
\alias{check}
\alias{print.fertile_check}
\title{Reproducbility checks}
\usage{
has_readme(path = ".", ...)

has_proj_root(path = ".", ...)

has_no_absolute_paths(path = ".", ...)

has_only_portable_paths(path = ".", ...)

has_no_randomness(path = ".", seed_old, ...)

check(path = ".", ...)

\method{print}{fertile_check}(x, ...)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{...}{currently ignore}

\item{seed_old}{The old seed before the code is rendered}

\item{x}{an object used to select a method.}
}
\description{
Reproducbility checks
}
