% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R, R/files.R
\name{checks}
\alias{checks}
\alias{check}
\alias{is_readme_exists}
\alias{has_proj_root}
\title{Reproducbility checks}
\usage{
check(path = ".", ...)

is_readme_exists(path = ".", ...)

has_proj_root(path = ".")
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{...}{currently ignore}
}
\description{
Reproducbility checks
}
