% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R, R/files.R, R/paths.R
\name{checks}
\alias{checks}
\alias{check_file}
\alias{file_exists_here}
\alias{check_file_here}
\alias{is_path_here}
\alias{check_path_here}
\alias{path_rel_here}
\title{Checking paths and files}
\usage{
check_file(file)

file_exists_here(path)

check_file_here(path, strict = TRUE)

is_path_here(path)

check_path_here(path, strict = TRUE)

path_rel_here(path)
}
\arguments{
\item{file}{a path to a file}

\item{path}{A vector of paths}
}
\value{
A logical vector

A logical vector
}
\description{
Checking paths and files

Does file exist in project tree?

Check if paths are relative to project root
}
\examples{
\dontrun{
check_file("data.csv")
check_file("~/.Rprofile")
check_file(tempdir())
}
file_exists_here(tempfile())
\dontrun{
check_file_here(tempfile())
}
is_path_here(tempfile())
}
