% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{danger}
\alias{danger}
\title{Override functions masked by fertile and run from the original packages.}
\usage{
danger(expr)
}
\arguments{
\item{expr}{Code to run as if \code{fertile} was not loaded}
}
\description{
Override functions masked by fertile and run from the original packages.
}
