% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_r_file}
\alias{is_r_file}
\title{Test whether a given path is to an R file}
\usage{
is_r_file(path)
}
\arguments{
\item{path}{Path to file you want to test}
}
\description{
Test whether a given path is to an R file
}
