% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_push}
\alias{log_push}
\title{Logging and reporting of file paths}
\usage{
log_push(x, .f, path = proj_root())
}
\arguments{
\item{x}{the path to capture}

\item{.f}{the function calling \code{x}}

\item{path}{Path to the \code{fertile} log file. Defaults to \code{\link{path_log}}}
}
\description{
Logging and reporting of file paths
}
\seealso{
\code{\link{shims}}
}
\keyword{internal}
