% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_push}
\alias{log_push}
\alias{log_report}
\alias{log_clear}
\alias{log_touch}
\title{Logging and reporting of file paths}
\usage{
log_push(x, .f)

log_report()

log_clear()

log_touch()
}
\arguments{
\item{x}{the path to capture}

\item{.f}{the function calling \code{x}}
}
\description{
These functions provide access to a log file, located at
\code{.fertile_paths.csv} in the project directory, that records the executed
paths passed to commonly-used input and output functions.
}
\examples{
log_report()
}
\seealso{
\code{\link{shims}}
}
