% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_report}
\alias{log_report}
\alias{log_clear}
\alias{log_touch}
\title{Reporting of logged file paths}
\usage{
log_report()

log_clear()

log_touch()
}
\description{
Reporting of logged file paths
}
\examples{
log_report()
}
