% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{proj_test}
\alias{proj_test}
\alias{proj_analyze}
\alias{proj_analyze_files}
\alias{proj_suggest_moves}
\alias{proj_move_files}
\alias{proj_analyze_pkgs}
\alias{proj_analyze_paths}
\title{Analyze project for reproducibility}
\usage{
proj_test(path = ".")

proj_analyze(path = ".")

proj_analyze_files(path = ".")

proj_suggest_moves(files)

proj_move_files(suggestions, execute = TRUE)

proj_analyze_pkgs(path = ".")

proj_analyze_paths(path = ".")
}
\arguments{
\item{path}{Path to project root}

\item{files}{List of files returned by \code{\link{proj_analyze_files}}}

\item{suggestions}{List of suggestsions returned by \code{\link{proj_suggest_moves}}}

\item{execute}{Do you want to actually move the files to their recommended location?}
}
\value{
A \code{fertile} object
}
\description{
Analyze project for reproducibility
}
\section{proj_test}{

Create a full report of project reproducibility. Includes:
packages referenced in code, files in the directory and suggestions for moving them,
and a list of paths that are not portable.

\code{proj_test("your project directory")}
}

\section{proj_analyze}{

Very similar to proj_test, except that this function does NOT
include a report of paths that are not portable.

\code{proj_analyze("your project directory")}
}

\section{proj_analyze_files}{

Provides a report of files present in a provided project directory.
Includes information about file size, extension, and a guess about the file type.

\code{proj_analyze_files("your project directory")}
}

\section{proj_suggest_moves}{

Takes a list of files returned by \code{\link{proj_analyze_files}}
and makes suggestions or where the files should be moved, as well
as a command for how to move them there.

\code{files <- proj_analyze_files("your project directory")}

\code{proj_suggest_moves(files)}
}

\section{proj_move_files}{

Execute the suggested commands returned by \code{\link{proj_suggest_moves}}.

\code{files <- proj_analyze_files("your project directory")}

\code{suggestions <- proj_suggest_moves(files)}

\code{proj_move_files(suggestions)}
}

\section{proj_analyze_pkgs}{

Returns all of the packages loaded in R code files as well as the
name of the files where they were referenced.

\code{proj_analyze_pkgs("your project directory")}
}

\section{proj_analyze_paths}{

Looks at paths used in R code located in a project directory and
reports paths that are absolute or that reference a location outside
the project directory.

\code{proj_analyze_paths("your project directory")}
}

