% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertile.R
\name{proj_test}
\alias{proj_test}
\alias{proj_analyze}
\alias{proj_render}
\alias{proj_report}
\alias{print.fertile}
\title{Analyze project for reproducibility}
\usage{
proj_test(path = ".")

proj_analyze(path = ".")

proj_render(path = ".")

proj_report(path = ".")

\method{print}{fertile}(x, ...)
}
\arguments{
\item{path}{Path to package root}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{fertile} object
}
\description{
Analyze project for reproducibility
}
\examples{
proj_test()
}
