% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sandbox}
\alias{sandbox}
\title{Utility function to create a copy of a project in a temp directory}
\usage{
sandbox(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Utility function to create a copy of a project in a temp directory
}
