% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{shims}
\alias{shims}
\alias{read_csv}
\alias{read.csv}
\alias{source}
\alias{write.csv}
\alias{write_csv}
\alias{setwd}
\alias{ggsave}
\title{Shims for common input/output functions}
\usage{
read_csv(file, ...)

read.csv(file, ...)

source(file, ...)

write.csv(x, file, ...)

write_csv(x, file, ...)

setwd(dir)

ggsave(filename, ...)
}
\arguments{
\item{file}{The path to the file}

\item{...}{arguments passed to functions}

\item{dir}{A character string: \link{tilde expansion} will be done.}

\item{filename}{File name to create on disk.}
}
\description{
Shims for common input/output functions
}
\seealso{
\code{\link[readr]{read_csv}}

\code{\link[utils]{read.csv}}

\code{\link[base]{source}}

\code{\link[utils]{write.csv}}

\code{\link[readr]{write_csv}}

\code{\link[base]{setwd}}

\code{\link[ggplot2]{ggsave}}
}
