% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{shims}
\alias{shims}
\alias{read_csv}
\alias{write_csv}
\alias{setwd}
\alias{source}
\title{Shims for common input/output functions}
\usage{
read_csv(file, ...)

write_csv(x, path, ...)

setwd(dir)

source(file, ...)
}
\arguments{
\item{file}{The path}

\item{...}{arguments passed to functions}

\item{x}{A data frame to write to disk}

\item{path}{Path or connection to write to.}

\item{dir}{A character string: \link{tilde expansion} will be done.}
}
\description{
Shims for common input/output functions
}
\examples{
\dontrun{
if (require(readr)) {
  read_csv(tempfile())
}
}
}
\seealso{
\code{\link[readr]{read_csv}}

\code{\link[readr]{write_csv}}

\code{\link[base]{setwd}}

\code{\link[base]{source}}
}
