% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{shims}
\alias{shims}
\alias{read.csv}
\alias{read.csv2}
\alias{read.delim}
\alias{read.delim2}
\alias{read.DIF}
\alias{read.fortran}
\alias{read.fwf}
\alias{read.table}
\alias{read_csv}
\alias{read_csv2}
\alias{read_delim}
\alias{read_file}
\alias{read_file_raw}
\alias{read_fwf}
\alias{read_lines}
\alias{read_lines_raw}
\alias{read_log}
\alias{read_table}
\alias{read_table2}
\alias{read_tsv}
\alias{read.dcf}
\alias{load}
\alias{source}
\alias{read_excel}
\alias{read.ftable}
\alias{fromJSON}
\alias{read.dta}
\alias{read.mtp}
\alias{read.spss}
\alias{read.systat}
\alias{read.sas7bdat}
\alias{tbl}
\alias{set.seed}
\alias{write.csv}
\alias{write_csv}
\alias{setwd}
\alias{save}
\alias{ggsave}
\alias{library}
\alias{require}
\title{Shims for common input/output functions}
\usage{
read.csv(file, ...)

read.csv2(file, ...)

read.delim(file, ...)

read.delim2(file, ...)

read.DIF(file, ...)

read.fortran(file, ...)

read.fwf(file, ...)

read.table(file, ...)

read_csv(file, ...)

read_csv2(file, ...)

read_delim(file, ...)

read_file(file, ...)

read_file_raw(file)

read_fwf(file, col_positions, ...)

read_lines(file, ...)

read_lines_raw(file, ...)

read_log(file, ...)

read_table(file, ...)

read_table2(file, ...)

read_tsv(file, ...)

read.dcf(file, ...)

load(file, envir = parent.frame(), verbose = FALSE)

source(file, ...)

read_excel(file, ...)

read.ftable(file, ...)

fromJSON(json_str, file, ...)

read.dta(file, ...)

read.mtp(file)

read.spss(file, ...)

read.systat(file, ...)

read.sas7bdat(file, ...)

tbl(src, ...)

set.seed(n, ...)

write.csv(x, file, ...)

write_csv(x, path, ...)

setwd(dir)

save(..., list = character(), file = stop("'file' must be specified"),
  ascii = FALSE, version = NULL, envir = parent.frame(),
  compress = isTRUE(!ascii), compression_level, eval.promises = TRUE,
  precheck = TRUE)

ggsave(filename, ...)

library(package, help, pos = 2, lib.loc = NULL,
  character.only = FALSE, logical.return = FALSE,
  warn.conflicts = TRUE, quietly = FALSE,
  verbose = getOption("verbose"))

require(package, lib.loc = NULL, quietly = FALSE,
  warn.conflicts = TRUE, character.only = FALSE)
}
\description{
Shims for common input/output functions
}
\seealso{
\code{\link[base]{library}}

\code{\link[base]{require}}
}
\keyword{internal}
