% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{shims}
\alias{shims}
\alias{read_csv}
\alias{read.csv}
\alias{load}
\alias{source}
\alias{tbl}
\alias{set.seed}
\alias{write.csv}
\alias{write_csv}
\alias{setwd}
\alias{save}
\alias{ggsave}
\alias{library}
\alias{require}
\title{Shims for common input/output functions}
\usage{
read_csv(file, ...)

read.csv(file, ...)

load(file, envir = parent.frame(), verbose = FALSE)

source(file, ...)

tbl(src, ...)

set.seed(n, ...)

write.csv(x, file, ...)

write_csv(x, path, ...)

setwd(dir)

save(..., list = character(), file = stop("'file' must be specified"),
  ascii = FALSE, version = NULL, envir = parent.frame(),
  compress = isTRUE(!ascii), compression_level, eval.promises = TRUE,
  precheck = TRUE)

ggsave(filename, ...)

library(package, help, pos = 2, lib.loc = NULL,
  character.only = FALSE, logical.return = FALSE,
  warn.conflicts = TRUE, quietly = FALSE,
  verbose = getOption("verbose"))

require(package, lib.loc = NULL, quietly = FALSE,
  warn.conflicts = TRUE, character.only = FALSE)
}
\description{
Shims for common input/output functions
}
\seealso{
\code{\link[base]{library}}

\code{\link[base]{require}}
}
\keyword{internal}
